LayoutTests/platform/qt-wk2/plugins/document-open-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: DOCUMENT OPEN SUCCESS
